Avtobot/requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.12.2
cachetools==4.2.2
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
lxml==5.1.0
python-telegram-bot==13.15
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
tinydb==4.8.0
tornado==6.1
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
